Podcast branch->trunk mergelog ------------------------------ this short documentation is intended for future reference
subjects
--------
branch: /contrexx/branches/contrexx_podcast (locally "contrexx_podcast")
trunk: /contrexx/trunk (locally "contrexx")
revisions merged: 1104:1141 
									1141:1142


How to find the BASE revision 
-----------------------------
see http://svnbook.red-bean.com/en/1.0/ch04s04.html


kon@kon539-srv:contrexx_podcast$ svn log --verbose --stop-on-copy
[...]
------------------------------------------------------------------------
r1104 | kon | 2007-07-12 17:54:26 +0200 (Thu, 12 Jul 2007) | 8 lines
Changed paths:
   A /branches/contrexx_podcast (from /podcast:1103)
[...]
------------------------------------------------------------------------
==>> BASE revision: 1104

change local working directory from branch working copy to the trunk working copy and merge the branch BASE revision into trunk HEAD:

kon@kon539-srv:contrexx_podcast$ cd ../contrexx #../contrexx is where a freshly checked out HEAD revision of the trunk lies
kon@kon539-srv:contrexx$ svn merge -r 1104:HEAD http://svn.contrexxlabs.com/contrexx/branches/contrexx_podcast
U    core_modules/news/admin.class.php
U    core_modules/news/template/module_news_settings.html
A    svn.sql.zip
U    lib/spamprotection/captcha.class.php
C    changelog/version_contrexx_v1.1.4/changelog.txt
U    admin/template/ascms/skins_examples.html
U    admin/index.php
U    core/skins.class.php
U    core/Navigation.class.php
U    core/Init.class.php
U    modules/forum/index.class.php
U    modules/forum/lib/forumLib.class.php
U    modules/forum/template/module_forum_settings.html
U    modules/forum/homeContent.class.php
C    modules/podcast/lang/de/backend.php
U    modules/podcast/index.class.php
C    modules/podcast/admin.class.php
U    modules/podcast/lib/podcastLib.class.php
U    modules/podcast/template/module_podcast_select_medium_source.html
U    modules/podcast/template/module_podcast_settings.html
C    modules/podcast/template/module_podcast_modify_medium.html
A    modules/podcast/homeContent.class.php
U    modules/immo/template/module_immo_add.html
U    index.php

dammit, forgot to commit latest contrexx_podcast changes! also merge that one into trunk:

kon@kon539-srv:contrexx$ svn merge -r 1141:1142 http://svn.contrexxlabs.com/contrexx/branches/contrexx_podcast
G    modules/podcast/index.class.php
C    modules/podcast/admin.class.php
G    modules/podcast/lib/podcastLib.class.php
U    modules/podcast/homeContent.class.php
G    index.php

a few conflicts occured, fortunately they were easy to solve.


git-svn-id: http://svn.contrexxlabs.com/contrexx/trunk@1143 5ce06e5d-2b28-0410-aa6a-b3cd9cf52d8a

+++ lib/spamprotection/captcha.class.php
@@ -19,7 +19,7 @@
  * @subpackage  lib_spamprotection
  */
 class Captcha {
-	
+
 	var $strRandomString;
 	var $strFontDir;
     var $strBackgroundDir;
@@ -33,9 +33,9 @@ class Captcha {
     */
     function Captcha() {
     	$this->__construct();
-    }    
-    
-    
+    }
+
+
 	/**
 	* Constructor: Initializes base-state of object
     *
@@ -46,14 +46,14 @@ class Captcha {
 	    $this->strBackgroundDir = ASCMS_DOCUMENT_ROOT.'/lib/spamprotection/backgrounds/';
 	    $this->strAbsPath 		= ASCMS_DOCUMENT_ROOT.'/images/spamprotection/';
 	    $this->strWebPath 		= ASCMS_PATH_OFFSET.'/images/spamprotection/';
-	    
+
 	    $this->cleanDirectory();
 	    $this->strRandomString 	= $this->createRandomString();
 	    $this->strFilename		= $this->createFilename($this->strRandomString);
 	    $this->createImage();
     }
 
-    
+
 	/**
 	 * Creates an random string with $intDigits digits.
 	 *
@@ -64,17 +64,17 @@ class Captcha {
     	if ($intDigits > 5) {
     		$intDigits = 5;
     	}
-    	
-		$strReturn = '';	    	
+
+		$strReturn = '';
         for ($i=1; $i <= $intDigits; ++$i) {
         	$intRand = rand(0,1);
         	$strReturn .= ($intRand %2 == 0) ? chr(rand(65,90)) : sprintf("%d", rand(2,9));
         }
-        
+
         return  $strReturn;
     }
-    
-    
+
+
     /**
      * Creates a new filename for the image.
      *
@@ -82,15 +82,15 @@ class Captcha {
      * @return	string		Created filename with the format ".jpg"
      */
     function createFilename($strRandomString) {
-    	
+
     	do {
     		$strFileName = md5(time().$strRandomString).'.jpg';
     	} while (is_file($this->strAbsPath.$strFileName));
-    	 
+
     	return $strFileName;
     }
-    
-    
+
+
 	/**
 	 * Creates the image into the temporary folder.
 	 *
@@ -98,32 +98,32 @@ class Captcha {
     function createImage() {
         $intBackground = rand(1, 7);
         $handleImage = imagecreatefromjpeg($this->strBackgroundDir.$intBackground.'.jpg');
-        
+
         $arrColors 	= array(	imagecolorallocate($handleImage, 0, 0, 0),		//black
         						imagecolorallocate($handleImage, 255, 0, 0), 	//red
         						imagecolorallocate($handleImage, 0, 180, 0), 	//darkgreen
         						imagecolorallocate($handleImage, 0, 105, 172),	//blue
         						imagecolorallocate($handleImage, 145, 19, 120)	//purple
         				);
-        				
+
         $arrFonts	= array(	$this->strFontDir.'coprgtb.ttf',
         						$this->strFontDir.'ltypeb.ttf',
         				);
-        
+
     	for ($i = 0; $i < strlen($this->strRandomString); ++$i) {
     		$intColor 	= rand(0,count($arrColors)-1);
     		$intFont	= rand(0,count($arrFonts)-1);
     		$intAngel 	= rand(-7,7);
     		$intYMove 	= rand(-4,+4);
-    		
+
 	        imagettftext($handleImage, 11, $intAngel,(1+$i*12), 20 + $intYMove, $arrColors[$intColor], $arrFonts[$intFont],substr($this->strRandomString,$i,1));
     	}
-        
+
         imagejpeg($handleImage, $this->strAbsPath.$this->strFilename,85);
         imagedestroy($handleImage);
     }
 
-    
+
     /**
      * Removes all images from the temporary folder, which are older than an hour.
      *
@@ -132,15 +132,15 @@ class Captcha {
 		$handleDir = opendir($this->strAbsPath);
 		if ($handleDir) {
 			while ($strFile = readdir($handleDir)) {
-				if ($strFile != '.' && $strFile != '..' && (filemtime($this->strAbsPath.$strFile) < time()-3600)) {
+				if (!is_dir($this->strAbsPath.$strFile) && $strFile != '.' && $strFile != '..' && (filemtime($this->strAbsPath.$strFile) < time()-3600)) {
 					unlink($this->strAbsPath.$strFile);
 				}
 			}
 			closedir($handleDir);
-		}    	
+		}
     }
-    
-    
+
+
 	/**
 	 * Creates an string which can be used as an alt tag for the protection-image.
 	 *
@@ -149,15 +149,15 @@ class Captcha {
     function getAlt() {
     	$strReturn = '';
     	$intLen = strlen($this->strRandomString);
-    	
+
     	for ($i = 0; $i < $intLen; ++$i) {
     		$strReturn .= substr($this->strRandomString,$i,1).'-';
     	}
-    	
+
         return substr($strReturn,0,-1);
 	}
 
-	
+
     /**
      * Return the relative URL to the created image.
      *
@@ -166,19 +166,19 @@ class Captcha {
     function getUrl() {
         return $this->strWebPath.$this->strFilename;
     }
-    
-    
+
+
     /**
      * Returns the md5-hash of the random number shown on the image. This function is received its name
-     * getOffset() for compatibility-reasons with older versions, which were offset-based. 
+     * getOffset() for compatibility-reasons with older versions, which were offset-based.
      *
      * @return	string		md5-hash of the random number
      */
     function getOffset() {
     	return md5($this->strRandomString);
     }
-    
-    
+
+
 	/**
 	 * This function is used to validate the entered values of a user.
 	 *
